+++ stubs/gtk4.php
@@ -2591,6 +2591,23 @@ class GdkPixbuf extends GObject
         unset($preserve_aspect_ratio);
         return null;
     }
+    /** Creates a new pixbuf by loading an image from an input stream. */
+    public static function new_from_stream(GInputStream $stream, ?GCancellable $cancellable = null): GdkPixbuf
+    {
+        unset($stream);
+        unset($cancellable);
+        return null;
+    }
+    /** Creates a new pixbuf by loading an image from an input stream. */
+    public static function new_from_stream_at_scale(GInputStream $stream, int $width, int $height, bool $preserve_aspect_ratio, ?GCancellable $cancellable = null): GdkPixbuf
+    {
+        unset($stream);
+        unset($width);
+        unset($height);
+        unset($preserve_aspect_ratio);
+        unset($cancellable);
+        return null;
+    }
     /**
      * Finishes an asynchronous pixbuf creation operation started with
      * gdk_pixbuf_new_from_stream_async().
@@ -2640,6 +2657,23 @@ class GdkPixbuf extends GObject
     {
         unset($path);
         return false;
+    }
+    /** Creates a new pixbuf by asynchronously loading an image from an input stream. */
+    public static function new_from_stream_async(GInputStream $stream, ?GCancellable $cancellable = null, ?callable $callback = null): void
+    {
+        unset($stream);
+        unset($cancellable);
+        unset($callback);
+    }
+    /** Creates a new pixbuf by asynchronously loading an image from an input stream. */
+    public static function new_from_stream_at_scale_async(GInputStream $stream, int $width, int $height, bool $preserve_aspect_ratio, ?GCancellable $cancellable = null, ?callable $callback = null): void
+    {
+        unset($stream);
+        unset($width);
+        unset($height);
+        unset($preserve_aspect_ratio);
+        unset($cancellable);
+        unset($callback);
     }
     /**
      * Finishes an asynchronous pixbuf save operation started with
@@ -2925,6 +2959,13 @@ class GdkPixbufAnimation extends GObject
         unset($resource_path);
         return null;
     }
+    /** Creates a new animation by loading it from an input stream. */
+    public static function new_from_stream(GInputStream $stream, ?GCancellable $cancellable = null): GdkPixbufAnimation
+    {
+        unset($stream);
+        unset($cancellable);
+        return null;
+    }
     /**
      * Finishes an asynchronous pixbuf animation creation operation started with
      * `new_from_stream_async`.
@@ -2933,6 +2974,13 @@ class GdkPixbufAnimation extends GObject
     {
         unset($async_result);
         return null;
+    }
+    /** Creates a new animation by asynchronously loading an image from an input stream. */
+    public static function new_from_stream_async(GInputStream $stream, ?GCancellable $cancellable = null, ?callable $callback = null): void
+    {
+        unset($stream);
+        unset($cancellable);
+        unset($callback);
     }
     /** Queries the height of the bounding box of a pixbuf animation. */
     public function get_height(): int
@@ -3933,6 +3981,129 @@ class GCancellable extends GObject
     }
 }
 /**
+ * `GInputStream` is a base class for implementing streaming input.
+ */
+class GInputStream extends GObject
+{
+    /** GInputStream is abstract in GTK: `new` only works on a PHP subclass (which gets its own GType). */
+    public function __construct()
+    {
+    }
+    /** Clears the pending flag on $stream. */
+    public function clear_pending(): void
+    {
+    }
+    /** Closes the stream, releasing resources related to it. */
+    public function close(?GCancellable $cancellable): bool
+    {
+        unset($cancellable);
+        return false;
+    }
+    /**
+     * Requests an asynchronous closes of the stream, releasing resources related to it. When the
+     * operation is finished $callback will be called. You can then call
+     * g_input_stream_close_finish() to get the result of the operation.
+     */
+    public function close_async(int $io_priority, ?GCancellable $cancellable, ?callable $callback): void
+    {
+        unset($io_priority);
+        unset($cancellable);
+        unset($callback);
+    }
+    /** Finishes closing a stream asynchronously, started from g_input_stream_close_async(). */
+    public function close_finish(GAsyncResult $result): bool
+    {
+        unset($result);
+        return false;
+    }
+    /** Checks if an input stream has pending actions. */
+    public function has_pending(): bool
+    {
+        return false;
+    }
+    /** Checks if an input stream is closed. */
+    public function is_closed(): bool
+    {
+        return false;
+    }
+    /**
+     * Finishes an asynchronous stream read operation started with g_input_stream_read_all_async().
+     */
+    public function read_all_finish(GAsyncResult $result): ?int
+    {
+        unset($result);
+        return null;
+    }
+    /**
+     * Like g_input_stream_read(), this tries to read $count bytes from the stream in a blocking
+     * fashion. However, rather than reading into a user-supplied buffer, this will create a new
+     * #GBytes containing the data that was read. This may be easier to use from language bindings.
+     */
+    public function read_bytes(int $count, ?GCancellable $cancellable): string
+    {
+        unset($count);
+        unset($cancellable);
+        return '';
+    }
+    /**
+     * Request an asynchronous read of $count bytes from the stream into a new #GBytes. When the
+     * operation is finished $callback will be called. You can then call
+     * g_input_stream_read_bytes_finish() to get the result of the operation.
+     */
+    public function read_bytes_async(int $count, int $io_priority, ?GCancellable $cancellable, ?callable $callback): void
+    {
+        unset($count);
+        unset($io_priority);
+        unset($cancellable);
+        unset($callback);
+    }
+    /** Finishes an asynchronous stream read-into-#GBytes operation. */
+    public function read_bytes_finish(GAsyncResult $result): string
+    {
+        unset($result);
+        return '';
+    }
+    /** Finishes an asynchronous stream read operation. */
+    public function read_finish(GAsyncResult $result): int
+    {
+        unset($result);
+        return 0;
+    }
+    /**
+     * Sets $stream to have actions pending. If the pending flag is already set or $stream is
+     * closed, it will return `false` and set $error.
+     */
+    public function set_pending(): bool
+    {
+        return false;
+    }
+    /** Tries to skip $count bytes from the stream. Will block during the operation. */
+    public function skip(int $count, ?GCancellable $cancellable): int
+    {
+        unset($count);
+        unset($cancellable);
+        return 0;
+    }
+    /**
+     * Request an asynchronous skip of $count bytes from the stream. When the operation is finished
+     * $callback will be called. You can then call g_input_stream_skip_finish() to get the result
+     * of the operation.
+     */
+    public function skip_async(int $count, int $io_priority, ?GCancellable $cancellable, ?callable $callback): void
+    {
+        unset($count);
+        unset($io_priority);
+        unset($cancellable);
+        unset($callback);
+    }
+    /** Finishes a stream skip operation. */
+    public function skip_finish(GAsyncResult $result): int
+    {
+        unset($result);
+        return 0;
+    }
+}
+/**
  * `GListModel` is an interface that represents a mutable list of `Object`. Its main intention is
  * as a model for various widgets in user interfaces, such as list views, but it can also be used
  * as a convenient method of returning lists of data, with support for updates.
@@ -4049,6 +4220,28 @@ class GListStore extends GObject implements GListModel
         unset($position);
         unset($removed);
         unset($added);
+    }
+}
+/**
+ * `GMemoryInputStream` is a class for using arbitrary memory chunks as input for GIO streaming
+ * input operations.
+ */
+class GMemoryInputStream extends GInputStream
+{
+    /** Creates a new empty #GMemoryInputStream. */
+    public function __construct()
+    {
+    }
+    /** Creates a new #GMemoryInputStream with data from the given $bytes. */
+    public static function new_from_bytes(string $bytes): GMemoryInputStream
+    {
+        unset($bytes);
+        return null;
+    }
+    /** Appends $bytes to data that can be read from the input stream. */
+    public function add_bytes(string $bytes): void
+    {
+        unset($bytes);
     }
 }
 /**
@@ -27162,6 +27355,12 @@ class WebKitWebView extends WebKitWebViewBase
         unset($save_mode);
         unset($cancellable);
         unset($callback);
+    }
+    /** Finish an asynchronous operation started with webkit_web_view_save(). */
+    public function save_finish(GAsyncResult $result): GInputStream
+    {
+        unset($result);
+        return null;
     }
     /** Asynchronously save the current web page. */
     public function save_to_file(string $file, WebKitSaveMode $save_mode, ?GCancellable $cancellable, ?callable $callback): void